Split app runtime policy from CLI
