+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "crates/hermes-core",
     "crates/hermes-provider-runtime",
+    "crates/hermes-app-runtime",
     "crates/hermes-agent",
     "crates/hermes-tools",
     "crates/hermes-gateway",
@@ -31,6 +32,7 @@ repository = "https://github.com/sheawinkler/hermes-agent-ultra"
 # Internal crates
 hermes-core = { path = "crates/hermes-core" }
 hermes-provider-runtime = { path = "crates/hermes-provider-runtime" }
+hermes-app-runtime = { path = "crates/hermes-app-runtime" }
 hermes-agent = { path = "crates/hermes-agent" }
 hermes-tools = { path = "crates/hermes-tools" }
 hermes-gateway = { path = "crates/hermes-gateway" }
