+++ crates/hermes-app-runtime/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "hermes-app-runtime"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Non-UI application runtime policy and agent configuration for Hermes"
+
+[dependencies]
+hermes-core = { workspace = true }
+hermes-config = { workspace = true }
+hermes-agent = { workspace = true }
+hermes-provider-runtime = { workspace = true }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3"
