+++ crates/hermes-cli/Cargo.toml
@@ -27,6 +27,7 @@ path = "src/main.rs"
 [dependencies]
 hermes-core = { workspace = true }
 hermes-provider-runtime = { workspace = true }
+hermes-app-runtime = { workspace = true }
 hermes-agent = { workspace = true }
 hermes-tools = { workspace = true }
 hermes-config = { workspace = true }
