+++ docs/parity/deep-problem-solving-diff.md
@@ -1,6 +1,6 @@
 # Deep Problem-Solving Diff
 
-Generated: `2026-06-22T09:01:07.499763+00:00`
+Generated: `2026-06-23T03:48:56.725733+00:00`
 
 ## Gate
