+++ docs/parity/global-parity-proof.json
@@ -260,7 +260,7 @@
       "unverified_cases": 0
     }
   },
-  "generated_at_utc": "2026-06-23T00:33:37.210464+00:00",
+  "generated_at_utc": "2026-06-23T03:48:57.156213+00:00",
   "gpar_completion": {
     "GPAR-01": true,
     "GPAR-02": true,
@@ -537,8 +537,8 @@
       "missing_rust_test_refs": 0,
       "queue_pending": 0,
       "queue_total": 6132,
-      "rust_test_files": 310,
-      "rust_test_functions": 3583,
+      "rust_test_files": 313,
+      "rust_test_functions": 3614,
       "test_intent_mapping_ratio": 1.0,
       "test_intents_mapped": 10,
       "test_intents_total": 10,
