+++ docs/parity/global-parity-proof.md
@@ -1,6 +1,6 @@
 # Global Parity Proof
 
-Generated: `2026-06-23T00:33:37.210464+00:00`
+Generated: `2026-06-23T03:48:57.156213+00:00`
 
 ## Gate Status
