+++ docs/releases/v0.16.0.md
@@ -48,7 +48,7 @@ git diff --check
 scripts/check-rust-runtime-no-python.sh
 scripts/check-runtime-placeholders.sh
 cargo test -p hermes-config agent_api_max_retries -- --nocapture
-cargo test -p hermes-cli test_build_agent_config_maps_agent_api_max_retries -- --nocapture
+cargo test -p hermes-app-runtime test_build_agent_config_maps_agent_api_max_retries -- --nocapture
 cargo test -p hermes-mcp stringified_mcp_arguments -- --nocapture
 cargo test -p hermes-mcp tools_call_coerces_stringified_object -- --nocapture
 cargo test -p hermes-cli run_dump_writes_real_saved_session_export_with_system_prompt -- --nocapture
